refactor(svelte-utils): split SessionGate (auth routing) and PersistenceGate (DI persistence boot)
The previous SessionGate conflated two boundaries: auth state routing AND
local persistence boot with recovery UI. That conflation forced every
signed-in payload to expose `whenReady` and `wipe` at the top level,
widening createSession's constraint with fields the session itself doesn't
need. The gate's UX was driving the data model from the wrong direction.

Split into two single-responsibility gates that compose:

  SessionGate         routes pending / signed-out / signed-in
                      props: session, signedIn, signedOut, pending?
                      no auth prop (session carries enough)

  PersistenceGate     awaits a load promise, offers Reload / Forget /
                      Sign out recovery
                      props: whenReady, wipe, auth?, children, pending?, error?
                      pure dependency injection: knows nothing about
                      bundle shape, IDB, sessions, or session payloads

createSession's constraint slims back to `{ userId } & Disposable`, which
is what the factory actually needs (user-switch detection + teardown).
The whenReady/wipe lines come out of every app's build() returns.

Each layout composes the two when both apply:

  <SessionGate {session}>
    {#snippet signedOut()}<AuthForm ... />{/snippet}
    {#snippet signedIn(s)}
      <PersistenceGate {auth} whenReady={s.x.idb.whenLoaded} wipe={...}>
        ...
      </PersistenceGate>
    {/snippet}
  </SessionGate>

Opensidian uses PersistenceGate alone (no auth state machine in its
layout, since its singleton client guarantees signed-in by mount time).
The 80-line inlined recovery UI from the previous commit goes away.

A future bundle without `idb`, with a multi-stage boot, or with no wipe
semantics still works: the caller injects whatever satisfies "a promise
to await" and "an action to call on forget device". The gate doesn't
encode bundle assumptions.
