+++ apps/fuji/src/lib/session.svelte.ts
@@ -42,8 +42,6 @@ export const session = createSession({
 					return deleted;
 				},
 			},
-			whenReady: fuji.idb.whenLoaded,
-			wipe: () => fuji.wipe(),
 			[Symbol.dispose]() {
 				entriesMap[Symbol.dispose]();
 				fuji[Symbol.dispose]();
