+++ apps/zhongwen/src/lib/session.svelte.ts
@@ -14,8 +14,6 @@ export const session = createSession({
 		return {
 			userId,
 			zhongwen,
-			whenReady: zhongwen.idb.whenLoaded,
-			wipe: () => zhongwen.wipe(),
 			[Symbol.dispose]() {
 				zhongwen[Symbol.dispose]();
 			},
