+++ packages/svelte-utils/src/persistence-gate/index.ts
@@ -1,0 +1,1 @@
+export { default as PersistenceGate } from './persistence-gate.svelte';
